-C and -x flags done
